Add authentication scripts.
The authentication setup is rather heavy at first glance. We want to
enable both quick and standard Jenkins users with usernames and
passwords as well as other authorization / authentication strategies
such as the GitHub-delegated authentication used on the ROS build farms
and the delegated authorization used on ci.ros2.org (admin access is
granted to members of a specific team).

The default strategy reads user information from a data bag
'ros_buildfarm_jenkins_users' and uses the Project-based Matrix
Authorization strategy to set up defined permissions for users and
enabling anonymous read access.

As it is we're using an attribute to determine whether this 'default'
mode should be used or a second strategy which expects to find a Groovy
script for configuring authentication/authorization in a data bag
'ros_buildfarm_jenkins_scripts'. This strategy is used when the
attribute is set to 'groovy'.