Restart jenkins if configuration file changes.
If this config file changes, we should restart Jenkins directly as
those changes may affect connect-able clients or cli tools which are
needed for the rest of the chef run.

This commit also adjusts the configuration path because as it turns out
the default path does not use the casc_configs directory but is in the
root of JENKINS_HOME, which is /var/lib/jenkins for the Debian package.