Use http for cli access.
The current documentation for the jenkins cookbook paints SSH keys as
the easy way to manage the local CLI but Jenkins's ssh access functions
differently than the http access in that it binds to an address and port
rather than a port only and due to the way Jenkins is configured this
seems like it's going to be the public interface rather than the
private one. They're one and the same on our instances so if the ssh
connection is opened on the local interface it will work but the
SSH port is randomized and advertized via the X-Jenkins-SSH header or
somesuch and the CLI will try to use the public IP or domain which
means it leaves the security group and tries to re-enter it but the port
is blocked.

Also http is just as easy, the bug we monkey patched in our ros2ci
cookbook is resolved in the upstream cookbook for 8.0.