Add publish over ssh config.
The publish-over-ssh plugin doesn't work with configuration as code.
The same missing API for c-as-c also makes it unlikely to work with a
Groovy script so we have to fall back to the the xml configuration.

This is largely borrowed from a live config and uses a data bag for the
host configurations.