Rework ssh publisher configuration.
As we work on the configuration I've been doing my best not to copy
verbatim patterns from the previous buildfarm_deployment but instead to
closely adapt those that make sense while losely adapting or entirely
revisiting things which either didn't translate well to chef or things
which have held back approachability or configurability in the past.

Let down by the publish-over-ssh plugin's lack of support for
configuration-as-code and groovy scripting I think I got a little too
abstract trying to clean up and deduplicate template logic.  The
"no-such-thing-as-two" pattern I follow suggested that the four
different targets we have for publish-over-ssh may someday be three or
five and therefore storing bulk host configurations as data bags made
sense in the moment. However when I looked at what that asked of our
community build farm operators I hesitated and ultimately made this
change.

While there are N similar configurations and N may grow or shrink in
time. Right now N=4 are required in order for the build farm to operate
as expected. One of the challenges with the puppet-based buildfarm
deployment is the lack of delineation between hiera config values which
must or may be changed by individual operators and which ones are
"static" to the build farm and are in hiera only because there did not
seem to be a convenient other mechanism to configure them.

I saw in this previous strategy that we would need to ask operators to
create data bags with static values in them and that's something I would
like to avoid if at all possible. Hardening on this point may require
review of existing work to check for violations but it's at least on my
mind going forward.

Thus in reconciling this most recent feature I have pushed much of the
configuration into attributes, using a data bag only for the secret
which should definitely be operator supplied. As a result I've committed
what I consider to be a somewhat grave act of metaprogramming for a chef
recipe in the interest of Don't-Repeat-Yourself which I'm not entirely
convinced is the overriding concern. I'm hoping that the current
implementation will serve until I find a better way to express the
intent while avoiding duplicate changes.

While making this change, I also realized that the method we're using to
override attributes related to the agent recipe may not work when we try
for the single-host approach which was too difficult to achieve with the
existing buildfarm_deployment but which is not impossible in theory
though it will require more configuration that we've done today. So I
noted some TODOs related to that aspect in this change.