Update nginx default server handling.
I accidentally put the default_server designation in the server name
where it becomes an extra server name instead of on the listen directive
where it works as intended. Since the default config on Debian installs
a separate default server that config file needs to be removed.

I also dropped the .conf extension from the rendered template as it's
not used for the Debian style nginx sites.