Rename data bag and expect environments.
The previous data bag name was easy to confuse with the actual host
private keys which we also store in data bags and this makes it clear
that we're using these keys to set known ssh hosts.

Since data bags aren't environment specific we have to find a way to use
different data for different environments. For these situations we're
going with using a single item with data for each environment within it.