Move generic kitchen config back to kitchen.yml.
So in my saga to get kitchen working reliably I tried a handful of
different kitchen drivers / providers with varying degreess of failure.

The primary issue for drivers other than vagrant and EC2 is the lack of
prebuilt boxes from the bento project.

Keeping all of the different kitchen configs around is going to cause
some confusion so I'd like to clean them up, but it's also possible to
have a single base kitchen.yml and override it with a kitchen.local.yml
so the plan going forward is to keep the general config in kitchen.yml
and only add changes from that default config to the
kitchen.$PROVIDER.yml files, one of which can be symlinked to
kitchen.local.yml for local testing.

We also no longer need a kitchen.vagrant.yml since Vagrant is the
default provider.