Install Podman for running contained pulp services.
The initial plan was to use systemd-docker but it appears to have
fallen out of maintenance and it's unclear if there is a successor
project with as much simplicity.

Podman's own systemd integration also proved to have caveats, largely
around the legibility of its generated systemd units which have names
like `container-OBNOXIOUSLYLONGCONTAINERID` instead of generating units
based on the container name. I ended up using the systemd unit @cottsay
and I wrote for systemd-docker as a starting point and folding in
Podman's techniques for managing container supervision. The result is
legible although not as neat as systemd-docker, now it just has to work,
which is what proved difficult for systemd-docker.

One difference between podman as we're using it and docker is that the
container is pre-created and in a ready state before being started in
the service unit whereas in the systemd-docker case the container is
created and removed by the unit when started and stopped.