Add gpg-vault service.
This is the adaptation of @cottsay's work creating an isolated gpg-vault
user running a shareable gpg-agent session.

Implementation note:
For other systemd services we've been installing the unit file and
triggering systemctl-daemon-reload manually. I noticed that chef has a
systemd_unit resource [1] and decided to try and use that to bundle
triggering the reload along with starting and enabling the service.

I'm interested in whether others like it better or not. It bakes systemd
into our chef but so does manually executing systemctl commands.
Feedback welcome.

[1]: https://docs.chef.io/resources/systemd_unit