Purge fingerprints directory on reconfiguration.
Credential files are recreated on each chef run in a non-idempontent
fashion. This may just be a quirk of the cookbook or it could be a
Jenkins limitation due to the per-instance encryption of sensitive data.
Either way we live with it for now by unconditionally purging the
fingerprints directory.