Add pulp client packages to agent.
These packages are used to connect to the pulp service.
There is no built in chef resource for python packages from pypi and the
options for python provider cookbooks seem quite stale in Chef
supermarket. With the use of exact versions the pip install command is
idempotent and pypi will uninstall any other versions of the package as
we update in the future.
As a result I think sticking with an execute resource rather than trying
to import a dependency is the way to go for now.