Add new 'group_execute' resource
The Chef 'execute' resource allows you to specify user and group
permissions with which to execute the command. Unfortunately, this
doesn't take into account seccondary groups listed in /etc/group.

We're currently using those seccondary groups to enable a user (process)
to access the GPG vault, and the reprepro commands invoked by chef will
need to perform some signing, and will therefore need permission.

The new resource defined by this change is nearly identical to the
'execute' resource, but adds an additional 'seccondary_groups' property.
The resource then augments the supplied command by prefixing it with
calls to 'sg <group-name>' to chain through the given groups. If a
'group' value was specified by the resource, that group is added to the
call chain last to ensure it is the primary group after the other groups
have been added.