Iterate over the group list backwards
Since the command is built from the inside-out to make the shell
escaping simpler, we want to start with the last group in the list,
which will then be switched to last in the call stack.

This should result in no functional change in behavior, but might make
the code a little bit easier to understand.