Restart jenkins-agent immediately on unit update
It appears that allowing the delayed restart to run immediately after
the service was started is causing it to fail.

This change issues a restart immediately after the unit file is created
or updated (and after systemd reloads it). In the case where the unit
was updated, the explicit 'start' is therefore a no-op. Otherwise, the
service is started at that time.