Enable caching of some of the RPM repository metadata
The names of files which match the caching rule contain the content
hashes, so the content at that URL should never change. These files
disappear as the repository is updated, and this leads to a race between
the loading of the repomd.xml and accessing the files listed within (and
here in this config).