Always force the renewing of the acme certificate with extra guards.
This command is designed to be idempotent itself, however any changes
made to the configuration must be made by forcing the renewal of the
certificate.

The guard is updated to check that `Le_ReloadCmd` is not empty because
the actual text of that guard is a base64 encoded string and while I've
reversed it once before I am not ready to do it again if I don't need
to.