Notify `gpg --import` resources when key files change.
This patch improves configuration behavior when key contents change.
When deploying changes to the public key, the keys on disk were updated
but the gpg keyring did not import the new keys since existing keys were
present.

Now, changes to the file resources will trigger runs of the import
execute resources which will provide the same behavior as before for
a freshly provisioned host but will correctly re-import keys when those
files change.