Create a separate copy of the public key on disk for pulp.
In order to avoid notifying the `pulp` user's gpg import when the
repository public key is changed, give the pulp user its own copy of
the key which is not changed until after the other pulp resources have
converged.

An alternative solution involves re-organizing the recipe so that the
necessary pulp infrastructure is created before the key file is changed
but this is a smaller, if sloppy, change.