Block potentially abusable publish-over-ssh connection test url.
This url is used by the connection test facility of the publish-over-ssh
plugin however this url is not properly authenticated and not restricted
to POST requests to prevent cross-site request forgery.

Since Jenkins is only publicly available via the nginx reverse proxy
this should allow us time to resolve these issues upstream or migrate
away from the plugin entirely.