Add ros_buildfarm.repo.server_name attribute.
I am bit conflicted about defaulting this to `nil`, the jenkins server
name defaults to `'ros_buildfarm'` and would biff if used with
letsencrypt enabled since it isn't a valid fqdn for the server name
verification.

`nil` breaks the convention of having a default but also gives us the
ability to raise the alarm early if a configuration isn't going to work,
rather than having the acme.sh certificate issue attempt fail.