Build the pulp docker image in its own directory.
The first thing that docker-build does is "send the build context to the
Docker daemon. This context includes the entire current directory of the
build. When rebuilding an image on an active host, this directory
includes the `media` subdirectory full of repository content.
On repo.ros2.org this directory is over 100GB in size.
As a result the docker build time is untenable. I have not even seen the
docker build start as I've aborted both builds after over 15 minutes of
"sending data".

The pulp_image dockerfile contains no ADD or COPY commands and so does
not appear to require any of the contents of the pulp data directory
beyond its own Dockerfile.