Add createrepo-agent to repo host alongside Pulp
This change adds the necessary configuration for using createrepo-agent
on the repo host for RPM repository management, and creates a separate
repository tree alongside the Pulp one for comparison.

When we're ready to switch, just drop the `_cra` suffix from the
repository directory name and remove all of the pulp stuff.