Make this weird case a fatal error rather.
I don't think this code is reachable in normal circumstances, Scott
pointed out that it could be a result of a change in CLI output or
localization differences in output. With that in mind, it makes more
sense to error in these cases since if the package is installed but at an unknowable version we should stop and make the user figure it out.