Add script to handle reprepro signing with multiple keys.
Reprepro will handle signing if an individual key is specified or if the
default gpg can be used, but signing a release with multiple keys
requires the use of a handler script.

The handler script is added as a static script which relies on a new
shell configuration file to set variables dictating the keys used via
adding them to KEY_OPTS.

Additional GPG options can also be set in the GPG_OPTS variable although
this is currently unused.

The config file is intended to be used as a chef template which will
incorporate the keys to be used.