Comment out de-listed plugin for Windows agent support.
This plugin is not used by the build farm, but it is an implicit
dependency of our current plugin set because it was extracted from the
Jenkins core and not all plugins we use have been updated since then.

Since we do not use Windows at all with the build farm, I don't believe
any functionality will be impacted.