Use subscribe to trigger gpg signature on rpm repo metadata.
This allows us to have multiple "trigger conditions" for updating the
repository metadata signatures manually. With the guard condition, we
would only create the signature if none currently existed, which did not
allow us to update it if we changed the signing key.

Use of `subscribes` rather than `notifies` is rarely preferred but in
this case, it makes the most sense because the subscribed-to resources
are finite whereas the notified resources are created in a loop based on
configuration attributes.